+++ dynamic_dynamodb/core/table.py
@@ -141,20 +141,21 @@ def __ensure_provisioning_reads(table_name, key_name, num_consec_read_checks):
     try:
         lookback_window_start = get_table_option(
             key_name, 'lookback_window_start')
+        lookback_period = get_table_option(key_name, 'lookback_period')
         current_read_units = dynamodb.get_provisioned_table_read_units(
             table_name)
         consumed_read_units_percent = \
             table_stats.get_consumed_read_units_percent(
-                table_name, lookback_window_start)
+                table_name, lookback_window_start, lookback_period)
         throttled_read_count = \
             table_stats.get_throttled_read_event_count(
-                table_name, lookback_window_start)
+                table_name, lookback_window_start, lookback_period)
         throttled_by_provisioned_read_percent = \
             table_stats.get_throttled_by_provisioned_read_event_percent(
-                table_name, lookback_window_start)
+                table_name, lookback_window_start, lookback_period)
         throttled_by_consumed_read_percent = \
             table_stats.get_throttled_by_consumed_read_percent(
-                table_name, lookback_window_start)
+                table_name, lookback_window_start, lookback_period)
         reads_upper_threshold = \
             get_table_option(key_name, 'reads_upper_threshold')
         reads_lower_threshold = \
@@ -195,6 +196,12 @@ def __ensure_provisioning_reads(table_name, key_name, num_consec_read_checks):
             get_table_option(key_name, 'increase_consumed_reads_with')
         increase_consumed_reads_scale = \
             get_table_option(key_name, 'increase_consumed_reads_scale')
+        decrease_consumed_reads_unit = \
+            get_table_option(key_name, 'decrease_consumed_reads_unit')
+        decrease_consumed_reads_with = \
+            get_table_option(key_name, 'decrease_consumed_reads_with')
+        decrease_consumed_reads_scale = \
+            get_table_option(key_name, 'decrease_consumed_reads_scale')
     except JSONResponseError:
         raise
     except BotoServerError:
@@ -392,31 +399,67 @@ def __ensure_provisioning_reads(table_name, key_name, num_consec_read_checks):
             updated_read_units = calculated_provisioning
 
     # Decrease needed due to low CU consumption
-    if (consumed_read_units_percent <= reads_lower_threshold
-            and not update_needed):
+    if not update_needed:
+        # If local/granular values not specified use global values
+        decrease_consumed_reads_unit = \
+            decrease_consumed_reads_unit or decrease_reads_unit
+
+        decrease_consumed_reads_with = \
+            decrease_consumed_reads_with or decrease_reads_with
+
+        # Initialise variables to store calculated provisioning
+        consumed_calculated_provisioning = scale_reader_decrease(
+            decrease_consumed_reads_scale,
+            consumed_read_units_percent)
+        calculated_provisioning = None
 
         # Exit if down scaling has been disabled
         if not get_table_option(key_name, 'enable_reads_down_scaling'):
             logger.debug(
                 '{0} - Down scaling event detected. No action taken as scaling'
-                'down reads has been disabled in the configuration'.format(
+                ' down reads has been disabled in the configuration'.format(
                     table_name))
         else:
-            if decrease_reads_unit == 'percent':
-                calculated_provisioning = calculators.decrease_reads_in_percent(
-                    updated_read_units,
-                    decrease_reads_with,
-                    get_table_option(key_name, 'min_provisioned_reads'),
-                    table_name)
-            else:
-                calculated_provisioning = calculators.decrease_reads_in_units(
-                    updated_read_units,
-                    decrease_reads_with,
-                    get_table_option(key_name, 'min_provisioned_reads'),
-                    table_name)
+            if consumed_calculated_provisioning:
+                if decrease_consumed_reads_unit == 'percent':
+                    calculated_provisioning = \
+                        calculators.decrease_reads_in_percent(
+                            updated_read_units,
+                            consumed_calculated_provisioning,
+                            get_table_option(
+                                key_name, 'min_provisioned_reads'),
+                            table_name)
+                else:
+                    calculated_provisioning = \
+                        calculators.decrease_reads_in_units(
+                            updated_read_units,
+                            consumed_calculated_provisioning,
+                            get_table_option(
+                                key_name, 'min_provisioned_reads'),
+                            table_name)
+            elif (reads_lower_threshold
+                  and consumed_read_units_percent > reads_lower_threshold
+                  and not decrease_consumed_reads_scale):
+                if decrease_consumed_reads_unit == 'percent':
+                    calculated_provisioning = \
+                        calculators.decrease_reads_in_percent(
+                            updated_read_units,
+                            decrease_consumed_reads_with,
+                            get_table_option(
+                                key_name, 'min_provisioned_reads'),
+                            table_name)
+                else:
+                    calculated_provisioning = \
+                        calculators.decrease_reads_in_units(
+                            updated_read_units,
+                            decrease_consumed_reads_with,
+                            get_table_option(
+                                key_name, 'min_provisioned_reads'),
+                            table_name)
 
-            if current_read_units != calculated_provisioning:
-                num_consec_read_checks = num_consec_read_checks + 1
+            if (calculated_provisioning
+                    and current_read_units != calculated_provisioning):
+                num_consec_read_checks += 1
 
                 if num_consec_read_checks >= num_read_checks_before_scale_down:
                     update_needed = True
@@ -439,6 +482,16 @@ def __ensure_provisioning_reads(table_name, key_name, num_consec_read_checks):
             logger.info(
                 '{0} - Increasing reads to meet min-provisioned-reads '
                 'limit ({1} reads)'.format(table_name, updated_read_units))
+
+    if calculators.is_consumed_over_proposed(
+            current_read_units,
+            updated_read_units,
+            consumed_read_units_percent):
+        update_needed = False
+        updated_read_units = current_read_units
+        logger.info(
+            '{0} - Consumed is over proposed read units. Will leave table at '
+            'current setting.'.format(table_name))
 
     logger.info('{0} - Consecutive read checks {1}/{2}'.format(
         table_name,
@@ -470,20 +523,21 @@ def __ensure_provisioning_writes(
     try:
         lookback_window_start = get_table_option(
             key_name, 'lookback_window_start')
+        lookback_period = get_table_option(key_name, 'lookback_period')
         current_write_units = dynamodb.get_provisioned_table_write_units(
             table_name)
         consumed_write_units_percent = \
             table_stats.get_consumed_write_units_percent(
-                table_name, lookback_window_start)
+                table_name, lookback_window_start, lookback_period)
         throttled_write_count = \
             table_stats.get_throttled_write_event_count(
-                table_name, lookback_window_start)
+                table_name, lookback_window_start, lookback_period)
         throttled_by_provisioned_write_percent = \
             table_stats.get_throttled_by_provisioned_write_event_percent(
-                table_name, lookback_window_start)
+                table_name, lookback_window_start, lookback_period)
         throttled_by_consumed_write_percent = \
             table_stats.get_throttled_by_consumed_write_percent(
-                table_name, lookback_window_start)
+                table_name, lookback_window_start, lookback_period)
         writes_upper_threshold = \
             get_table_option(key_name, 'writes_upper_threshold')
         writes_lower_threshold = \
@@ -524,6 +578,12 @@ def __ensure_provisioning_writes(
             get_table_option(key_name, 'increase_consumed_writes_with')
         increase_consumed_writes_scale = \
             get_table_option(key_name, 'increase_consumed_writes_scale')
+        decrease_consumed_writes_unit = \
+            get_table_option(key_name, 'decrease_consumed_writes_unit')
+        decrease_consumed_writes_with = \
+            get_table_option(key_name, 'decrease_consumed_writes_with')
+        decrease_consumed_writes_scale = \
+            get_table_option(key_name, 'decrease_consumed_writes_scale')
     except JSONResponseError:
         raise
     except BotoServerError:
@@ -532,7 +592,8 @@ def __ensure_provisioning_writes(
     # Set the updated units to the current read unit value
     updated_write_units = current_write_units
 
-    # Reset consecutive write count num_write_checks_reset_percent is reached
+    # Reset consecutive write count if num_write_checks_reset_percent
+    # is reached
     if num_write_checks_reset_percent:
 
         if consumed_write_units_percent >= num_write_checks_reset_percent:
@@ -723,35 +784,70 @@ def __ensure_provisioning_writes(
             updated_write_units = calculated_provisioning
 
     # Decrease needed due to low CU consumption
-    if (consumed_write_units_percent
-            <= writes_lower_threshold and not update_needed):
+    if not update_needed:
+        # If local/granular values not specified use global values
+        decrease_consumed_writes_unit = \
+            decrease_consumed_writes_unit or decrease_writes_unit
 
-        # Exit if up scaling has been disabled
+        decrease_consumed_writes_with = \
+            decrease_consumed_writes_with or decrease_writes_with
+
+        # Initialise variables to store calculated provisioning
+        consumed_calculated_provisioning = scale_reader_decrease(
+            decrease_consumed_writes_scale,
+            consumed_write_units_percent)
+        calculated_provisioning = None
+
+        # Exit if down scaling has been disabled
         if not get_table_option(key_name, 'enable_writes_down_scaling'):
             logger.debug(
-                '{0} - Down scaling event detected. No action taken as scaling '
-                'down writes has been disabled in the configuration'.format(
+                '{0} - Down scaling event detected. No action taken as scaling'
+                ' down writes has been disabled in the configuration'.format(
                     table_name))
         else:
-            if decrease_writes_unit == 'percent':
-                calculated_provisioning = \
-                    calculators.decrease_writes_in_percent(
-                        current_write_units,
-                        decrease_writes_with,
-                        get_table_option(key_name, 'min_provisioned_writes'),
-                        table_name)
-            else:
-                calculated_provisioning = calculators.decrease_writes_in_units(
-                    current_write_units,
-                    decrease_writes_with,
-                    get_table_option(key_name, 'min_provisioned_writes'),
-                    table_name)
+            if consumed_calculated_provisioning:
+                if decrease_consumed_writes_unit == 'percent':
+                    calculated_provisioning = \
+                        calculators.decrease_writes_in_percent(
+                            updated_write_units,
+                            consumed_calculated_provisioning,
+                            get_table_option(
+                                key_name, 'min_provisioned_writes'),
+                            table_name)
+                else:
+                    calculated_provisioning = \
+                        calculators.decrease_writes_in_units(
+                            updated_write_units,
+                            consumed_calculated_provisioning,
+                            get_table_option(
+                                key_name, 'min_provisioned_writes'),
+                            table_name)
+            elif (writes_lower_threshold
+                  and consumed_write_units_percent > writes_lower_threshold
+                  and not decrease_consumed_writes_scale):
+                if decrease_consumed_writes_unit == 'percent':
+                    calculated_provisioning = \
+                        calculators.decrease_writes_in_percent(
+                            updated_write_units,
+                            decrease_consumed_writes_with,
+                            get_table_option(
+                                key_name, 'min_provisioned_writes'),
+                            table_name)
+                else:
+                    calculated_provisioning = \
+                        calculators.decrease_writes_in_units(
+                            updated_write_units,
+                            decrease_consumed_writes_with,
+                            get_table_option(
+                                key_name, 'min_provisioned_writes'),
+                            table_name)
 
-            if current_write_units != calculated_provisioning:
-                num_consec_write_checks = num_consec_write_checks + 1
+            if (calculated_provisioning and
+                    current_write_units != calculated_provisioning):
+                num_consec_write_checks += 1
 
-                if (num_consec_write_checks >=
-                        num_write_checks_before_scale_down):
+                if num_consec_write_checks >= \
+                        num_write_checks_before_scale_down:
                     update_needed = True
                     updated_write_units = calculated_provisioning
 
@@ -772,6 +868,16 @@ def __ensure_provisioning_writes(
             logger.info(
                 '{0} - Increasing writes to meet min-provisioned-writes '
                 'limit ({1} writes)'.format(table_name, updated_write_units))
+
+    if calculators.is_consumed_over_proposed(
+            current_write_units,
+            updated_write_units,
+            consumed_write_units_percent):
+        update_needed = False
+        updated_write_units = current_write_units
+        logger.info(
+            '{0} - Consumed is over proposed write units. Will leave table at '
+            'current setting.'.format(table_name))
 
     logger.info('{0} - Consecutive write checks {1}/{2}'.format(
         table_name,
@@ -842,10 +948,12 @@ def __ensure_provisioning_alarm(table_name, key_name):
     """
     lookback_window_start = get_table_option(
         key_name, 'lookback_window_start')
+    lookback_period = get_table_option(key_name, 'lookback_period')
+
     consumed_read_units_percent = table_stats.get_consumed_read_units_percent(
-        table_name, lookback_window_start)
+        table_name, lookback_window_start, lookback_period)
     consumed_write_units_percent = table_stats.get_consumed_write_units_percent(
-        table_name, lookback_window_start)
+        table_name, lookback_window_start, lookback_period)
 
     reads_upper_alarm_threshold = \
         get_table_option(key_name, 'reads-upper-alarm-threshold')
@@ -859,8 +967,7 @@ def __ensure_provisioning_alarm(table_name, key_name):
     # Check upper alarm thresholds
     upper_alert_triggered = False
     upper_alert_message = []
-    if (reads_upper_alarm_threshold > 0 and
-            consumed_read_units_percent >= reads_upper_alarm_threshold):
+    if 0 < reads_upper_alarm_threshold <= consumed_read_units_percent:
         upper_alert_triggered = True
         upper_alert_message.append(
             '{0} - Consumed Read Capacity {1:f}% '
@@ -870,8 +977,7 @@ def __ensure_provisioning_alarm(table_name, key_name):
                 consumed_read_units_percent,
                 reads_upper_alarm_threshold))
 
-    if (writes_upper_alarm_threshold > 0 and
-            consumed_write_units_percent >= writes_upper_alarm_threshold):
+    if 0 < writes_upper_alarm_threshold <= consumed_write_units_percent:
         upper_alert_triggered = True
         upper_alert_message.append(
             '{0} - Consumed Write Capacity {1:f}% '
@@ -944,6 +1050,28 @@ def scale_reader(provision_increase_scale, current_value):
                 return scale_value
             else:
                 scale_value = provision_increase_scale.get(limits)
+        return scale_value
+    else:
+        return scale_value
+
+
+def scale_reader_decrease(provision_decrease_scale, current_value):
+    """
+
+    :type provision_decrease_scale: dict
+    :param provision_decrease_scale: dictionary with key being the
+        scaling threshold and value being scaling amount
+    :type current_value: float
+    :param current_value: the current consumed units or throttled events
+    :returns: (int) The amount to scale provisioning by
+    """
+    scale_value = 0
+    if provision_decrease_scale:
+        for limits in sorted(provision_decrease_scale.keys(), reverse=True):
+            if current_value > limits:
+                return scale_value
+            else:
+                scale_value = provision_decrease_scale.get(limits)
         return scale_value
     else:
         return scale_value